r24577 | kobus | 2019-11-21 11:06:32 -0700 (Thu, 21 Nov 2019) | 5 lines
Made it so that for mac OS, which "ar" we use is first searched for wherever the
c compiler is. This is because using the gcc ar with clang does not (always)
work.

Recorded result of commands processed by svn in /home/ivilab/repo_sync/ivilab_fake_delphi_local/src/Make:
Updating '.':
U    init_compile
Updated to revision 24577.